Fixed another bug.
git-svn-id: http://svn.code.sf.net/p/greensql/code/trunk/greensql-fw@116 7e47f64f-3c34-0410-9091-801b9fe948be

+++ src/proxymap.cpp
@@ -43,7 +43,7 @@ void wrap_Server(int fd, short which, void * arg)
 {
     long proxy_id;
     
-    proxy_id = (int) arg;
+    proxy_id = (long) arg;
     logevent(NET_DEBUG, "[%d]wrap_Server\n", proxy_id);
     GreenSQLConfig * conf = GreenSQLConfig::getInstance();
